requirements.txt
python-dotenv
requests
pandas
sqlalchemy
psycopg2-binary
pytz
loguru
